peripherals: add a PeripheralSet struct that lets us join together a set of things that together implement the full set of peripheral traits
It's becoming more and more apparent that we don't really need the Peripherals supertrait. It'd be just fine to have Control trait implementations be generic over all the peripheral traits and leave it at that; we don't really need to enforce that the Control implementator has/is using a full set of peripherals.
The supertrait is actually a bit of a pain since we'll need to go implement each of the peripheral traits for the PeripheralSet container (blanket impls, but still -- irritating) so that Peripherals, the supertrait, can be implemented for the PeripheralSet container.

It does buy us a little in the way of ergonomics though (the interpreter can be generic over some P: Peripherals instead of G, A, P, T, C, I, O) so I'm not going to ditch it yet.