sim: checkpoint: splitting up step and adding some macros for readability
there's I! and i!: I! is more ambitious and more magic than i! (though not as much as I'd like; defining a macro in the macro didn't seem to be enough to pick up the regs from the local scope so you still need to do R[reg])

Leaving both in for now since I'm not entirely confident I! actually does the right thing, always.

This commit contains much carnage that's being committed for documentation purposes; it'll disappear in a future clean up commit.