traits: attempt another feature rebalance
Instead of "no_std and std_functionality; pick one" it's now just no_std: enabled or disabled.
I made this change because it _seems_ to make it possible to actually have users of lc3_traits in the workspace that want no_std and don't want no_std to coexist.

I'm still not super clear on what Cargo does with features and workspaces but it seems to, for a particular dependency, take the union of all the features every crate in the workspace wants for that dependency and then proceed to compile that dependency for the workspace with that union of features. Hopefully if a particular feature is explicitly enabled and explicitly disabled for different crates in the workspace, Cargo will do something more sensible (we'll see).