sim: fix a bug / get HALT to actually work
We set the MCR runnin bit in the `InstructionInterpreter::halt` impl so that the state is consistent (in case halt is not triggered from LC-3 userspace). The MCR device also calls this function when it's written to in a way that causes a halt. This recurses infinitely (or at least until the stack overflows).
We already check (in halt) that we're not already halted before we write to the MCR: this is actually sufficient to prevent the stack overflow *if* we (in the MCR write method) write the bit _before_ we call `InstructionInterpreter::halt`.

The odd bit here is that `InstructionInterpreter::halt` only exists so that the MCR device can change the state of the interpreter to `Halted`; it should *never* be called by anyone but the MCR device. Unfortuntely it's part of a 'public' interface since we have to make it part of the trait, so we'll still handle this case.