sim: update to use EventFuture
It makes sense to use SharedStateState since it has basically the shape you'd want for this kind of thing (i.e. signal to an external thing that we've hit an event) and it's already written.
I started writing SimFuture to use EventFuture and it immediately turned into a literal copy of EventFuture and that's how we got here.