control: tweak the Transport trait to provide a mutable reference to self to implementors
We've only got one impl of Transport at the moment (MPSC) and this doesn't require (non-interior) mutation, but it's very likely other impls of Transport will need this (i.e. UART + a FIFO for the TM4C); hence this change.
The users of Transport that we've got (Device and Controller) currently have ownership of their Transport implementors anyways, so this shouldn't cause any real problems.