control: fix encoding by not using the Encoding trait
This was a mistake. Though the (blanket) impl for Encoding ensures properties like <Self as Encode<_>>::Encoding == <Self as Decode<_>>::Encoding the trait itself does not enforce this (it cannot since we cannot have bounds on associated types without GATs; a past commit contains an attempt to use another type parameter on Encoding in lieu of an Encoded type paramter but that led to recursive evaluation related overflows so I ditched it pretty quickly). So, I ended up needing to add, for example, an A: Encode<_, Encoded = B> bound in addition to the Encoding<_> bound (since, again, though our impl guarentees such things anyways, the trait doesn't) which should have worked but didn't (errors about the type not being able to be inferred (?)).
So, I ditched the associated types on Encoding. Best not to duplicate if we can't provide the guarentees we want for the duplicates.

Encoding is now basically a marker trait with some functions that will disappear in a future commit.