peripherals: add set/get ProgramMetadata to the Memory trait
Not 100% convinced this is the place for it but it seems to make sense. Partly because Memory already will have to deal with peristent storage on platforms and is set up to vary per platform (unlike, for example, the Simulator, which is the other place where I think this could possibly go).