macros: add a derive proc macro to impl Display using Debug
Could have used enum_display_derive or derive_more for this but this was quick and this is a good start to writing the macro that we want.

Notes within about what said macro will eventually need to do.

Later, this should probably not delegate to Debug since it's (afaict) v hard or impossible for us to check that a type actually has a Debug impl within the proc macro meaning that we'll never be able to give good errors for that case. Instead we should generate the Display impl for the situations we're aiming to support (enums with variants w/o associated data).