control: add a progress api to load
This is slightly over-engineered (i.e. the traits, one of which is private - because it contains functions that should not be user accessible - essentially making it sealed).
I realized we'd need Sync/interior mutability for these to be useful, but it did not, at first, occur to me that all we need are atomics. I thought we'd need to do Arc<Mutex<T>> style wrappers + blanket impls like we did with SharedStateState and the Peripheral shims which is why the traits even exist (same kinda deal as SharedStateState: no_std compatible type that isn't Sync with Sync wrappers for it on std platforms).

We can clean it up later. Or not. It's probably fine.