tui: add primitive backoff support to the Device Thread
This is untuned and also is currently not exponential.

Our situation is a little weird because the downside of erring on the side of sleeping for too long isn't just lag/a decrease in responsivity for the user; instead it also means that the controller thread will consume massive amounts of CPU since it's effectively spinning waiting on the response.
So, this is a somewhat elaborate balancing act.

re: #52