sim: move the test infra back out of the crate
Unfortunately, things marked with cfg(test) inside a crate don't seem to be visible to integration tests (I think the cfg(test) doesn't propagate to the crate being tested in the same way that dev-dependencies in tests aren't (I think) aware they're being used in a test suite).

The right move is to have a non-publishable test-infrastructure crate in this workspace, but this will do for now even if it is a hack.