control: update Controller with the Enc/Dec changes
Unforunately because we have to receive/send messages for every Control call and because not all the functions on Control provide a mutable reference to self (for good reason!) I had to stick the Enc/Dec instances in a RefCell. I don't think this will end up being a problem in practice (RefCells are marked as !Sync) but it's still a little gross.