shims: compromise on the lifetimes
After many battles with `rustc`, we've arrived at exactly where we were 2 months ago.

We have to choose to either have the interrupt lifetime be larger than the inner
source/sink lifetime (`'int: 'io`) or the other way around (`'io: 'int`). In either case,
the actual peripheral trait impl is only valid for the interrupt lifetime (since the trait
has the interrupt lifetime as a type parameter).

The latter case was desirable for this very narrow and specific use case: interrupt flags
on the stack, static (unused) sources and sinks. However, this raises a question: what does
it mean to implement a peripheral trait for the lifetime of 'int when some of the fields
of the type you're implementing the trait on live for 'io where 'io lives longer than 'int?
Or vice versa: if 'int: 'io what does it mean for the trait to be implemented for 'int
(longer than all the fields on the type actually are valid for)?

So, actually, I think this is the wrong way to think about it; the bounds on the type
parameters of the blanket impl (i.e. the lifetimes 'int and 'io) end up deciding for which
lifetime combination applies. The actual impl of the trait is not really aware of 'io at all;
so long as the type it's implemented on is alive (note: the lifetime of the type is
implicitly bounded by 'io; the "whichever of these two lifetimes is smaller" bound that we
wanted is already _there_), the trait implementation is happy.

This is to say, the lifetimes should be unbounded in the blanket impl. With the `'int: 'io`
bound in this commit, we end up requiring 'int to live for 'static in the test-infra macro.
When we elide 'io I think it gets _inferred_ to be 'static which is why that doesn't help.