app-support: lifetime notes/changes
Now that the lifetimes on the InputShim/OutputShim actually work right we can drop some
unnecessary lifetime bounds!

I gave the Box::leak mess another attempt but I think this is as good as we can do for now.

If we switch to having the BlackBox actually just allocate a particular device's necessary
storage (obviating the need for the downcast-to-Any stuff which requires 'static we can get
rid of Box::leak.

Another way is to do something spooky like hold a witness for the lifetime on BlackBox and
then do some unsafe crimes to _turn the thing into a 'static_ (probably by Box::leaking)
and making sure that the reference we give out only lives for as long as the witness.
Problems with this include that we can't actually drop the thing at the right time (i.e.
when the lifetime runs out) which is actually fine. More importantly that Box::leak
(correctly!) doesn't actually give us back a 'static ref if the thing isn't static to begin
with! So this wouldn't actually work. :-/

Not sure if it's even possible to transmute to a 'static ref without doing lots of UB (as in
even if we could guarantee that no one _else_ sees a 'static ref the fact that we'd use it
to downcast to Any is probably already UB).

This is workable for now though, I think. The test runner is probably the only thing that
spins up lots of instances of the simulator (i.e it's the only thing for which Box::leak
would actually be a problem) and that uses its own infrastructure instead of `Init` anyways.

So, to anyone reading this, if this ends up mattering to you, the associated type thing is
probably the way to go. Future problems for the future!