control: update Encode to take a reference to the message
We want this so we can plug the hole that having to impl Clone for LoadApiSeesion exposes.

An unfortunate side effect of this is that it means that the Transparent impl doesn't work anymore unless the type implements Clone. This is only a 'problem' for the sim-rpc (running locally with rpc) option that we have. In the next commit, there'll be a potential fix for this.