control: update the Transport trait's return types
Pretty sure there's a better way to actually do this; the point is that we have three different states:
  - got a message
  - didn't get a message, but no errors
  - got an error while trying to get a message

As the comments in this commit discuss, not sure if Result<Option<M>, Err> or Result<M, Option<Err>> expresses this better. Ultimately I think the ergonomics for both are suboptimal. I think the ideal solution is to have an enum with the above three variants _and_ to impl Try for it. Once Try goes stable we can switch over, I guess. (TODO)