shims: tweaks/fixes for the I/O shims
The Input thing is pretty clearly a bug and there's work we can/should do in the future to clean up the interface and the InputShim impl (duplicated state, confusing/extraneous functions).

The Output thing is less clear. Ultimately I do think that the OutpuShim shouldn't do this kind of checking (it should just siphon bytes), but the wrinkle is Unicode support. Our current plan for unicode support would require this kind of checking in the OutputShim and also since this peripheral is supposed to be a Display, one could argue that the intent is for the bytes transferred to be output to a screen. As in: it may be reasonable for us to assume that the things transmitted via it are characters/text.
For now though, dropping the ASCII check should be okay.