app-support: expose the wasm timeout function (when on wasm)
This kind of thing is definitely a little icky (as in, the event_loop module doesn't have all that much to do with this function and this kind of thing should really go in `util` or it's own crate). Another valid concern is that there isn't a not_wasm equivalent of this function.

Even so, I think this is small, documented, and not that egregious so let's call it okay for now.